Add installation instructions back to documentation homepage and consolidate instances across docs.
These instructions were removed in #17744, but it has been suggested
that it would be useful to have easy access to this info directly from
the docs homepage. Adding this directly would result in too many copies
of the installation instructions scattered throughout the docs, so this
PR creates a new stub page to hold the installation instructions, which
is then `include`-ed in all the relevant places.